+++ test/components/toast/toast.e2e-spec.js
@@ -35,10 +35,10 @@ describe('Toast example-index tests', () => {
     await buttonEl.click();
 
     expect(await element.all(by.id('toast-container')).count()).toEqual(1);
-    await browser.driver.wait(protractor.ExpectedConditions.visibilityOf(await element(by.id('toast-container'))), config.waitsFor);
+    await browser.driver.sleep(config.sleep);
 
     await element(by.css('#toast-container button.btn-close')).click();
-    await browser.driver.wait(protractor.ExpectedConditions.invisibilityOf(await element(by.id('toast-container'))), config.waitsFor);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.id('toast-container')).isDisplayed()).toBeFalsy();
   });